+++ projects/ui-kit/src/lib/ui-kit/table/components/table-config-aside/rt-table-config-aside.component.ts
@@ -5,7 +5,8 @@ import { MatIconButton } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
 import { MatTooltip } from '@angular/material/tooltip';
 
-import { BlockDirective, ElemDirective, Nullable } from '@rt-tools/core';
+import { BlockDirective, ElemDirective } from '@rt-tools/core';
+import { Nullable } from '@rt-tools/utils';
 import { areArraysEqual } from '@rt-tools/utils';
 import { BreakpointService, RtIconOutlinedDirective } from '@rt-tools/core';
 import { ASIDE_REF, AsideRef } from '../../../aside/aside.types';