+++ projects/ui-kit/src/lib/ui-kit/table/util/table-component.interface.ts
@@ -1,7 +1,7 @@
 import { InjectionToken, Signal, TemplateRef, WritableSignal } from '@angular/core';
 import { MatFormFieldAppearance } from '@angular/material/form-field';
 
-import { Nullable } from '@rt-tools/core';
+import { Nullable } from '@rt-tools/utils';
 import { FilterModel, SortModel } from '@rt-tools/utils';
 import { ITable } from './table-column.interface';